Make port numbers greater than 999 work in all locales.
Based on fix provided by David J.

lexical_cast converts a number to a string in a locale-dependent way.  For example 65535 may be converted to "65,535" or "65.535" depending on the user's locale.  When converting port numbers to strings for the purposes of using the string in a moniker, the port number must be canonical.  In other words, it must be locale-independent.

This change replaces lexical_cast with a custom conversion that forces a particular locale (C classic) regardless of the user's locale.