Add file renaming to libssh2 wrapper and use to implement Provider.
This also fixes a bug we weren't aware of: previously, if the server supported SFTP 5+ and the RENAME_ATOMIC flag, we wouldn't prompt the user to confirm overwrite if the file already existed.  It's not clear why we ever thought RENAME_ATOMIC prevented overwriting, but we did.  Fortunately OpenSSH servers don't support the flag anyway as they use SFTP 3.