Moved libssh2 SFTP exception-wrapper functions into `ssh::detail::libssh2` namespace.
This is the first stage of a refactoring that will eventually have all libssh2 exception wrappers in the `ssh::detail::libssh2` namespace.  This helps us define a boundary for the yet-to-be-implemented thread locking.  Any functions in the `ssh::detail::libssh2` namespace must not perform any locking but the session must be locked before they are called.  This ensure the exception throw definitely correspond to the function called.

Also removed the requirement that the pointers be supplied as `shared_ptr`s.  These functions should be usable however we finally decide to manage the pointers, which may well not be using a `shared_ptr` as they make object lifetime difficult to control.