Remove accessors to `sftp_channel` private members.
The `sftp_channel` class was exposing its privates publicly via two accessor methods, `session` and `get`.  Obviously bad.

This commit removes those accessors.  A side effect of this is that most SFTP filesystem operations become methods of `sftp_channel` as they need access to the channels private members.  Really they always should have been methods as they are intrinsic to the nature of a filesystem, not higher-level operations implemented in terms of what a filesystem offers.

A similar side effect is that instances of `directory_iterator` are now created via `sftp_channel`.  This is because the iterator implements intrinsic operations of the channel so it makes sense that the channel creates the iterator as an extension of itself rather than the iterator taking the channel as a parameter.  The latter would imply that the iterator operated on the public interface of the channel, which it does not.

The same change should be extended to the SFTP stream classes but that remains for a future commit.  For the moment, they are made friends so that can access the channel's privates.