Fix keyboard-interactive authentication error reporting.
We forgot to pass the error code to `ssh::detail::libssh2::userauth::keyboard_interactive_ex`, so failures to authenticate were being thrown as exceptions rather than prompting the user to try again (we think).  Also we were passing a session parameter to `translate_status` that we no longer need to process errors.