Fix non-thread-safe SFTP object destruction.
There was a bug in our session locking where the session could be locked during an operation on one thread and meanwhile another thread was destroying a libssh2 file handle or SFTP channel.  The destructors weren't locking the session so they would not wait for the operation that had locked the thread to complete.  This would corrupt the session and lead to all kinds of nastyness.

This change fixes the problem by giving file handles and SFTP channels their own RAII classes that gracefully handle their shutdown (instead of relying on `shared_ptr` which makes it less clear what's happening) and lock the session before doing so.