Make SSH classes non-copyable to better reflect their underlying concepts.
SSH sessions, for instance, are not inherently copyable so making the class that represents the copyable is somewhat misleading.  The most important flaw was that a session is not shut down until all _copies_ of the object are destroyed, making it very hard to guarantee a session has been disconnected.

This change fixes that by making the objects non-copyable to their underlying session/channel/handle is destroyed at the same time as the representing object.